Split `metanome_run.py` into multiple files
* Also fixes an oversight in `file_column_statistics`
