+++ pysrc/models/metanome_run.py
@@ -1,111 +1,10 @@
-import datetime
-import json
-import math
-from math import isnan
-import os
 from dataclasses import dataclass
 from dataclasses_json import dataclass_json
-import statistics
-from typing import Iterator
 
-from pysrc.models.column_information import ColumnInformation
-from pysrc.models.errors import ErrorMetric, INDType, MissingValues
-from pysrc.models.ind import IND
 from pysrc.models.column_statistics import ColumnStatistic
+from pysrc.models.metanome_run_configuration import MetanomeRunConfiguration
+from pysrc.models.metanome_run_results import MetanomeRunResults
 from pysrc.utils.dataclass_json import DataclassJson
-
-from pysrc.utils.descriptive_statistics import file_column_statistics
-from pysrc.utils.ind_credibility import ind_credibility
-
-
-@dataclass_json
-@dataclass(frozen=True)
-class MetanomeRunConfiguration(DataclassJson):
-    """Contains configuration information about a Metanome run"""
-    algorithm: str
-    arity: str
-    total_budget: int | None
-    sampling_method: str
-    allowed_missing_values: int
-    time: datetime.datetime
-
-    source_dir: str
-    source_files: list[str]
-    tmp_folder: str
-    results_folder: str
-    result_suffix: str
-    output_folder: str
-    clip_output: bool
-    header: bool
-    print_inds: bool
-    create_plots: bool
-
-    is_baseline: bool
-
-    def __hash__(self) -> int:
-        return hash((self.algorithm, self.arity, self.total_budget, self.sampling_method, self.allowed_missing_values, self.time, self.source_dir,
-                     tuple(self.source_files), self.tmp_folder, self.results_folder,
-                     self.result_suffix, self.output_folder, self.clip_output, self.header,
-                     self.print_inds, self.create_plots, self.is_baseline))
-
-    def credibility(self) -> float:
-        """Get the credibility (i.e. how trustworthy this config is) of the config."""
-        # TODO: Actually depend this on the config data
-        # return float(product([budget for budget in self.total_budget]))
-        if self.total_budget is None:
-            raise ValueError(self.total_budget)
-        return self.total_budget
-
-@dataclass_json
-@dataclass(frozen=True)
-class MetanomeRunResults(DataclassJson):
-    inds: list[IND]
-
-    def has_ind(self, other_ind: IND) -> bool:
-        """This checks whether this object has an IND that is identical to the passed-in one,
-        i.e. whether they are from the same table and column, but may differ by name"""
-        # Check whether it's directly contained
-        if other_ind in self.inds:
-            return True
-        clean_other_ind = IND(
-            dependents=[
-                ColumnInformation(table_name=column.table_name.split('__')[0], column_name=column.column_name)
-                for column
-                in other_ind.dependents
-            ], referenced=[
-                ColumnInformation(table_name=column.table_name.split('__')[0], column_name=column.column_name)
-                for column
-                in other_ind.referenced
-            ])
-        # Check whether cleaned other is directly contained
-        if clean_other_ind in self.inds:
-            return True
-
-        clean_inds = [
-            IND(
-                dependents=[
-                    ColumnInformation(table_name=column.table_name.split('__')[0], column_name=column.column_name)
-                    for column
-                    in ind.dependents
-                ], referenced=[
-                    ColumnInformation(table_name=column.table_name.split('__')[0], column_name=column.column_name)
-                    for column
-                    in ind.referenced
-                ])
-            for ind
-            in self.inds
-        ]
-        # Check whether cleaned version is in cleaned version
-        return clean_other_ind in clean_inds
-
-    def __len__(self) -> int:
-        return len(self.inds)
-
-    def __iter__(self) -> Iterator[IND]:
-        return self.inds.__iter__()
-
-    def __hash__(self) -> int:
-        return hash((tuple(self.inds)))
 
 
 @dataclass_json
@@ -114,331 +13,3 @@ class MetanomeRun(DataclassJson):
     configuration: MetanomeRunConfiguration
     column_statistics: list[ColumnStatistic]
     results: MetanomeRunResults
-
-
-@dataclass_json
-@dataclass(frozen=True)
-class MetanomeRunBatch(DataclassJson):
-    runs: list[MetanomeRun]
-
-    def __len__(self) -> int:
-        return len(self.runs)
-
-    def __iter__(self) -> Iterator[MetanomeRun]:
-        return self.runs.__iter__()
-
-    @property
-    def baseline(self) -> MetanomeRun:
-        return next(run for run in self.runs if run.configuration.is_baseline)
-
-    def ranked_inds(self) -> dict[IND, float]:
-        # Collect INDs
-        ind_map: dict[tuple[str, str], IND] = {}  # Map from (dependent, referenced) -> IND
-        inds: dict[IND, list[tuple[int, MetanomeRun]]] = {}
-        for run in self.runs:
-            for ind in run.results.inds:
-                if (str(ind.dependents), str(ind.referenced)) not in ind_map:
-                    clean_ind = IND(dependents=ind.dependents, referenced=ind.referenced)
-                    ind_map[(str(ind.dependents), str(ind.referenced))] = clean_ind
-                    inds[clean_ind] = []
-
-        for run in self.runs:
-            # Skip baseline (it won't be available in production)
-            # TODO: Is this correct behavior? Not sure, but otherwise we'll get *every* TP IND as we're checking the baseline.
-            if run.configuration.is_baseline:
-                continue
-            for ind in run.results.inds:
-                clean_ind = ind_map[(str(ind.dependents), str(ind.referenced))]
-                # if clean_ind not in inds: inds[clean_ind] = []
-                inds[clean_ind].append((ind.missing_values('dict'), run))
-        # maximum_missing_values = max(missing_values for configMissingValuePairs in inds.values() for missing_values, _ in configMissingValuePairs)
-        baseline = self.baseline
-        inds_credibilities = {
-                ind: [
-                    ind_credibility(ind, run, missing_values, baseline)
-                    for missing_values, run
-                    in configMissingValuesPairs
-                    ]
-                for ind, configMissingValuesPairs
-                in inds.items()
-            }
-        for ind in inds_credibilities:
-            if len(inds_credibilities[ind]) > 0:
-                continue
-            # When this IND was not found in the runs, set -2.
-            inds_credibilities[ind] = [-2.0]
-        # Rank INDs by SUM over ALL runs (with value 0.0 for runs that didn't find the IND)
-        ranked_inds = { ind: credibility_sum if not isnan(credibility_sum := sum(credibilities)) else -1.0 for ind, credibilities in inds_credibilities.items() }
-        return ranked_inds
-
-
-
-def parse_results(result_file_name: str, *, algorithm: str, arity: str, results_folder: str, print_inds: bool,
-                  is_baseline: bool, header: bool) -> MetanomeRunResults:
-    """Parses result file and returns run results"""
-    ind_list: list[IND] = []
-    lines: list[str] = []
-    try:
-        with open(os.path.join(os.getcwd(), results_folder, result_file_name), 'r') as file:
-            lines = file.readlines()
-    except FileNotFoundError:
-        return MetanomeRunResults(ind_list)
-
-    for line in lines:
-        line_json = json.loads(line)
-        errors: list[ErrorMetric] = []
-        if arity == 'unary' and is_baseline == True:
-            dependant_raw = line_json['dependant']['columnIdentifiers'][0]
-            dependant_table = dependant_raw['tableIdentifier'].rsplit('.', 1)[0]
-            dependant_column = dependant_raw['columnIdentifier']
-            dependant = ColumnInformation(table_name=dependant_table, column_name=dependant_column)
-
-            referenced_raw = line_json['referenced']['columnIdentifiers'][0]
-            referenced_table = referenced_raw['tableIdentifier'].rsplit('.', 1)[0]
-            referenced_column = referenced_raw['columnIdentifier']
-            referenced = ColumnInformation(table_name=referenced_table, column_name=referenced_column)
-
-            if algorithm == 'PartialSPIDER':
-                missing_values = line_json["missingValues"]
-                errors.append(MissingValues(missing_values))
-            # TODO: Figure out better way to identify inds. Is this parsing even necessary?
-            ind = IND(dependents=[dependant], referenced=[referenced], errors=errors)
-
-        elif arity == 'unary' and is_baseline == False:
-            dependant_raw = line_json['dependant']['columnIdentifiers'][0]
-            dependant_table = dependant_raw['tableIdentifier'].rsplit('.', 1)[0].split('__', 1)[0]
-            dependant_column = dependant_raw['columnIdentifier'] if header else 'column' + str(dependant_raw['tableIdentifier'].rsplit('.', 1)[0].rsplit('_')[-1])
-            dependant = ColumnInformation(table_name=dependant_table, column_name=dependant_column)
-
-            referenced_raw = line_json['referenced']['columnIdentifiers'][0]
-            referenced_table = referenced_raw['tableIdentifier'].rsplit('.', 1)[0].split('__', 1)[0]
-            referenced_column = referenced_raw['columnIdentifier'] if header else 'column' + str(referenced_raw['tableIdentifier'].rsplit('.', 1)[0].rsplit('_')[-1])
-            referenced = ColumnInformation(table_name=referenced_table, column_name=referenced_column)
-
-            if algorithm == 'PartialSPIDER':
-                missing_values = line_json["missingValues"]
-                errors.append(MissingValues(missing_values))
-            # TODO: Figure out better way to identify inds. Is this parsing even necessary?
-            ind = IND(dependents=[dependant], referenced=[referenced], errors=errors)
-
-        elif arity == 'nary' and is_baseline == True:
-            dependant_list: list[ColumnInformation] = []
-            dependant_raw = line_json['dependant']['columnIdentifiers']
-            for dependant_entry in dependant_raw:
-                dependant_table = dependant_entry['tableIdentifier'].rsplit('.', 1)[0]
-                dependant_column = dependant_entry['columnIdentifier']
-                dependant = ColumnInformation(table_name=dependant_table, column_name=dependant_column)
-                dependant_list.append(dependant)
-
-            referenced_list: list[ColumnInformation] = []
-            referenced_raw = line_json['referenced']['columnIdentifiers']
-            for referenced_entry in referenced_raw:
-                referenced_table = referenced_entry['tableIdentifier'].rsplit('.', 1)[0]
-                referenced_column = referenced_entry['columnIdentifier']
-                referenced = ColumnInformation(table_name=referenced_table, column_name=referenced_column)
-                referenced_list.append(referenced)
-
-            ind = IND(dependents=dependant_list, referenced=referenced_list)
-
-        elif arity == 'nary' and is_baseline == False:
-            dependant_list = []
-            dependant_raw = line_json['dependant']['columnIdentifiers']
-            for dependant_entry in dependant_raw:
-                dependant_table = dependant_entry['tableIdentifier'].rsplit('.', 1)[0].split('__', 1)[0]
-                dependant_column = dependant_entry['columnIdentifier'] if header else 'column' + str(dependant_entry['tableIdentifier'].rsplit('.', 1)[0].rsplit('_')[-1])
-                dependant = ColumnInformation(table_name=dependant_table, column_name=dependant_column)
-                dependant_list.append(dependant)
-
-            referenced_list = []
-            referenced_raw = line_json['referenced']['columnIdentifiers']
-            for referenced_entry in referenced_raw:
-                referenced_table = referenced_entry['tableIdentifier'].rsplit('.', 1)[0].split('_', 1)[0]
-                referenced_column = referenced_entry['columnIdentifier'] if header else 'column' + str(referenced_entry['tableIdentifier'].rsplit('.', 1)[0].rsplit('_')[-1])
-
-                referenced = ColumnInformation(table_name=referenced_table, column_name=referenced_column)
-                referenced_list.append(referenced)
-
-            ind = IND(dependents=dependant_list, referenced=referenced_list)
-        else:
-            continue
-
-        ind_list.append(ind)
-
-    return MetanomeRunResults(ind_list)
-
-
-def run_metanome(configuration: MetanomeRunConfiguration, output_fname: str, pipe: bool) -> MetanomeRun:
-    # TODO: Make these configurable
-    if configuration.algorithm == 'BINDER':
-        algorithm_path = 'BINDER.jar'
-        algorithm_class_name = 'de.metanome.algorithms.binder.BINDERFile'
-    elif configuration.algorithm == 'PartialSPIDER':
-        algorithm_path = 'PartialSPIDER.jar'
-        algorithm_class_name = 'de.metanome.algorithms.spider.SPIDERFile'
-    else:
-        raise ValueError(configuration.algorithm)
-
-    metanome_cli_path = 'metanome-cli.jar'
-    separator = '\\;'
-    escape = '\\\\'
-    output_rule = f'file:{output_fname}'
-    allowed_gb: int = 6
-
-    # Calculate File Statistics
-    source_files_column_statistics = [stats for f in configuration.source_files for stats in file_column_statistics(f, header=configuration.header, is_baseline=configuration.is_baseline)]
-
-    # Construct Command
-    file_name_list = ' '.join([f'"{file_name}"' for file_name in configuration.source_files])
-
-    execute_str = f'java -Xmx{allowed_gb}g -cp {metanome_cli_path}:{algorithm_path} de.metanome.cli.App \
-                    --algorithm {algorithm_class_name} \
-                    --files {file_name_list} \
-                    --separator {separator} \
-                    --file-key INPUT_FILES \
-                    --skip-differing-lines \
-                    -o {output_rule} \
-                    --escape {escape} '
-
-    if configuration.header:
-        execute_str += '--header '
-
-    if configuration.algorithm == 'BINDER':
-        execute_str += f'--algorithm-config DETECT_NARY:{"true" if configuration.arity == "nary" else "false"}'
-    else:
-        execute_str += '--algorithm-config TEMP_FOLDER_PATH:SPIDER_temp,\
-                        MAX_MEMORY_USAGE_PERCENTAGE:60,\
-                        INPUT_ROW_LIMIT:-1,\
-                        CLEAN_TEMP:true,\
-                        MEMORY_CHECK_FREQUENCY:100'
-        if not configuration.is_baseline:
-            execute_str += f',MAX_NUMBER_MISSING_VALUES:{configuration.allowed_missing_values}'
-        else:
-            execute_str += ',MAX_NUMBER_MISSING_VALUES:0'
-        
-    if pipe:
-        execute_str += ' | tail -n 0'
-    elif configuration.clip_output:
-        execute_str += ' | tail -n 2'
-
-    # Run
-    os.system(execute_str)
-    # Parse
-    result = parse_results(result_file_name=output_fname + configuration.result_suffix,
-                           algorithm=configuration.algorithm,
-                           arity=configuration.arity,
-                           results_folder=configuration.results_folder,
-                           print_inds=configuration.print_inds,
-                           is_baseline=configuration.is_baseline,
-                           header=configuration.header)
-    return MetanomeRun(configuration=configuration, column_statistics=source_files_column_statistics, results=result)
-
-
-# For unary INDs, this method returns absolute counts for TP, FP, FN, etc.
-def compare_csv_line_unary(inds: list[IND], baseline: MetanomeRunResults):
-    tp, fp = 0, 0
-    sum_tp_missing_values, sum_fp_missing_values = 0, 0
-    num_inds = len(inds)
-
-    for ind in inds:
-        if baseline.has_ind(ind):
-            ind.errors.append(INDType('TP'))
-            tp += 1
-            sum_tp_missing_values += ind.missing_values('dict')
-        else:
-            ind.errors.append(INDType('FP'))
-            fp += 1
-            sum_fp_missing_values += ind.missing_values('dict')
-
-    fn = len(baseline.inds) - tp
-
-    if num_inds > 0:
-        precision = tp / (tp + fp) if tp + fp != 0 else float('nan')
-        recall = tp / (tp + fn) if tp + fn != 0 else float('nan')
-        f1 = 2 * (precision * recall) / (precision + recall) if recall + precision != 0 else float('nan')
-        mean_tp_missing_values = sum_tp_missing_values / tp if tp > 0 else float('nan')
-        mean_fp_missing_values = sum_fp_missing_values / fp if fp > 0 else float('nan')
-    else:
-        precision, recall, f1, mean_tp_missing_values, mean_fp_missing_values = 0, 0, 0, 0, 0
-
-    return tp, fp, fn, precision, recall, f1, mean_tp_missing_values, mean_fp_missing_values
-
-
-# For nary INDs, this returns lists with counts for each arity
-def compare_csv_line_nary(inds: list[IND], baseline: MetanomeRunResults):
-    max_arity = max([ind.arity() for ind in baseline.inds])
-
-    tp, fp = [0 for _ in range(max_arity)], [0 for _ in range(max_arity)]
-    inds_per_arity = [0 for _ in range(max_arity)]
-    for ind in baseline.inds:
-        inds_per_arity[ind.arity() - 1] += 1
-
-    for ind in inds:
-        arity = ind.arity() - 1  # -1 to match list indices
-        if baseline.has_ind(ind):
-            ind.errors.append(INDType('TP'))
-            tp[arity] += 1
-        else:
-            ind.errors.append(INDType('FP'))
-            fp[arity] += 1
-
-    fn = [inds_per_arity[arity] - tp[arity] for arity in range(max_arity)]
-
-    precision, recall, f1 = [0.0 for _ in range(max_arity)], [0.0 for _ in range(max_arity)], [0.0 for _ in
-                                                                                               range(max_arity)]
-    for i in range(max_arity):
-        if tp[i] + fp[i] > 0:
-            precision[i] = tp[i] / (tp[i] + fp[i])
-
-        if tp[i] + fn[i] > 0:
-            recall[i] = tp[i] / (tp[i] + fn[i])
-
-        if recall[i] + precision[i] > 0:
-            f1[i] = 2 * (precision[i] * recall[i]) / (precision[i] + recall[i])
-        else:
-            f1[i] = float('nan')
-
-    return tp, fp, fn, precision, recall, f1
-
-
-def run_as_compared_csv_line(run: MetanomeRun, baseline: MetanomeRunResults) -> list[str]:
-    sampled_file_paths = run.configuration.source_files
-    sampled_file_names = [path.rsplit('/', 1)[-1].replace('.csv', '') for path in sampled_file_paths]
-
-    file_names, methods, budgets = [], [], []
-    for sampled_file in sampled_file_names:
-        split_filename = sampled_file.split(
-            '__')  # Detect Column Sampling, as this is evident from the '__' in the file
-        split_metadata = []
-        if len(split_filename) == 2:
-            split_metadata = split_filename[1].split('_')
-        split_filename = [split_filename[0]]
-
-        if len(split_metadata) == 3:
-            split_filename.append(split_metadata[0])
-            split_filename.append(split_metadata[1])
-        if len(split_filename) == 3:
-            fname, budget, sampling_method = split_filename
-            fname = fname + '_' + split_metadata[2]
-        else:
-            fname, budget, sampling_method = sampled_file, str(float('inf')), 'None'
-
-        file_names.append(fname)
-        methods.append(sampling_method)
-        budgets.append(budget)
-
-    if run.configuration.arity == 'unary':
-        tp, fp, fn, precision, recall, f1, mean_tp_missing_values, mean_fp_missing_values = compare_csv_line_unary(run.results.inds, baseline)
-        return ['; '.join(file_names), '; '.join(methods), '; '.join(budgets), str(tp), str(fp), str(fn), f'{precision:.3f}', f'{recall:.3f}', f'{f1:.3f}', f'{mean_tp_missing_values:.3f}', f'{mean_fp_missing_values:.3f}']
-
-    else:
-        tp, fp, fn, precision, recall, f1 = compare_csv_line_nary(run.results.inds, baseline)
-
-        return ['; '.join(file_names),
-                '; '.join(methods),\
-                '; '.join(budgets), \
-                '; '.join([str(tp_i) for tp_i in tp]), \
-                '; '.join([str(fp_i) for fp_i in fp]), \
-                '; '.join([str(fn_i) for fn_i in fn]), \
-                '; '.join([f'{precision_i:.3f}' for precision_i in precision]), \
-                '; '.join([f'{recall_i:.3f}' for recall_i in recall]), \
-                '; '.join([f'{f1_i:.3f}' for f1_i in f1])]
